Fixes issue #51
The getFields() function can return a field name like "a",
and if that field is an array of embedded documents, it will
return something like "a.0.b" too. This way the $f[$field]
can either be undefined, int(1), or an array. The original
check fails to recognize the second condition. This proposed
modification fixes this issue.